+++ src/tab_bar.rs
@@ -21,10 +21,12 @@ use std::any::Any;
 use std::cmp::max;
 use std::cmp::min;
 use std::isize;
+use std::mem::replace;
 
 use gui::Cap;
 use gui::ChainEvent;
 use gui::Event;
+use gui::EventChain;
 use gui::Handleable;
 use gui::Id;
 use gui::Key;
@@ -32,6 +34,7 @@ use gui::UiEvent;
 use gui::UiEvents;
 
 use event::EventUpdated;
+use in_out::InOut;
 use selection::SelectionState as SelectionStateT;
 use state::State;
 use task_list_box::TaskListBox;
@@ -40,6 +43,62 @@ use termui::TermUiEvent;
 
 /// The selection state as used by a `TabBar`.
 pub type SelectionState = SelectionStateT<Id>;
+
+
+/// An enum representing the states a search can be in.
+#[derive(Debug, PartialEq)]
+enum SearchT<T>
+where
+  T: PartialEq,
+{
+  /// No search is currently in progress.
+  Unset,
+  /// A search was started but the state is currently being transferred
+  /// to another widget.
+  // Note that in principle we could use the Unset variant to fudge this
+  // case, but we would like to assert different invariants when the
+  // state is taken versus when it was never actually set.
+  Taken,
+  /// The full state of a search. The first value is the (lowercased)
+  /// text to search for, the second one represents the task that was
+  /// selected last, and the third one represents the selection state.
+  State(String, SearchState, T),
+}
+
+impl<T> SearchT<T>
+where
+  T: PartialEq,
+{
+  /// Take the value, leave the `Taken` variant in its place.
+  fn take(&mut self) -> SearchT<T> {
+    replace(self, SearchT::Taken)
+  }
+}
+
+/// The selection state as used by a `TabBar`.
+type Search = SearchT<SelectionState>;
+
+
+/// An enum capturing the search behavior on an individual tab.
+#[derive(Debug, PartialEq)]
+pub enum SearchState {
+  /// Start the search at the currently selected task.
+  Current,
+  /// Start the search at the first task in the query being displayed.
+  First,
+  /// Start the search at the task with the given index.
+  Task(usize),
+}
+
+impl SearchState {
+  /// Check if the state has the `First` variant active.
+  fn is_first(&self) -> bool {
+    match self {
+      SearchState::First => true,
+      _ => false,
+    }
+  }
+}
 
 
 /// Sanitize a selection index.
@@ -55,6 +114,7 @@ pub struct TabBar {
   tabs: Vec<(String, Id)>,
   selection: usize,
   prev_selection: usize,
+  search: Search,
 }
 
 impl TabBar {
@@ -86,6 +146,54 @@ impl TabBar {
       tabs: tabs,
       selection: selection,
       prev_selection: selection,
+      search: SearchT::Unset,
+    }
+  }
+
+  /// Handle a `TermUiEvent::SearchTask` event.
+  fn handle_search_task(&mut self,
+                        string: String,
+                        search_state: SearchState,
+                        mut selection_state: SelectionState) -> Option<UiEvents> {
+    match self.search {
+      SearchT::Taken => {
+        // We have to distinguish three cases here, in order:
+        // 1) No matching task was to be found on all tabs, i.e., we
+        //    wrapped around already. In this case we should display an
+        //    error.
+        // 2) No task matching the search string was found on the tab we
+        //    just issued the search to, in which case we need to try
+        //    the next tab.
+        // 3) The next task in line was found and selected, in which
+        //    case we just store the search state and wait for
+        //    additional user input to select the next one or similar.
+        if selection_state.has_cycled() {
+          self.search = SearchT::State(string.clone(), search_state, selection_state);
+
+          let error = format!("Text '{}' not found", string).to_string();
+          let event = TermUiEvent::SetInOut(InOut::Error(error));
+          Some(UiEvent::Custom(Box::new(event)).into())
+        } else {
+          selection_state.reset_cycled();
+
+          if selection_state.has_advanced() {
+            debug_assert!(search_state.is_first());
+
+            let iter = self.tabs.iter().map(|x| x.1);
+            let new_idx = selection_state.normalize(iter);
+            let tab = self.tabs.get(new_idx).unwrap().1;
+
+            let event = TermUiEvent::SearchTask(string, SearchState::First, selection_state);
+            let event = UiEvent::Returnable(self.id, tab, Box::new(event));
+            Some(ChainEvent::Event(event))
+          } else {
+            self.search = SearchT::State(string, search_state, selection_state);
+            None
+          }
+        }
+      },
+      SearchT::Unset |
+      SearchT::State(..) => panic!("invalid search state"),
     }
   }
 
@@ -93,7 +201,11 @@ impl TabBar {
   fn handle_custom_event(&mut self,
                          event: Box<TermUiEvent>,
                          cap: &mut dyn Cap) -> Option<UiEvents> {
-    match *event {
+    // See Rust issue #16223: We cannot properly move out of an enum
+    // where one part of a tuple inside does not implement Copy. This
+    // screws up the `SearchTask` case. Use identity function trick here
+    // to work around this problem.
+    match { *event } {
       TermUiEvent::SelectTask(task_id, mut state) => {
         let iter = self.tabs.iter().map(|x| x.1);
         let new_idx = state.normalize(iter);
@@ -112,7 +224,32 @@ impl TabBar {
         let update = self.set_select(select as isize, cap);
         (None as Option<Event>).maybe_update(update)
       },
-      _ => Some(UiEvent::Custom(event).into()),
+      TermUiEvent::EnteredText(mut string) => {
+        if !string.is_empty() && !self.tabs.is_empty() {
+          string.make_ascii_lowercase();
+
+          let _ = self.search.take();
+          let tab = self.selected_tab();
+          let state = SelectionState::new(tab);
+
+          let event1 = TermUiEvent::SetInOut(InOut::Search(string.clone()));
+          let event1 = UiEvent::Custom(Box::new(event1));
+
+          let event2 = TermUiEvent::SearchTask(string, SearchState::Current, state);
+          let event2 = UiEvent::Returnable(self.id, tab, Box::new(event2));
+
+          Some(UiEvents::from(event1).chain(event2)).update()
+        } else {
+          None
+        }
+      },
+      TermUiEvent::InputCanceled => None,
+      TermUiEvent::SearchTask(string, search_state, selection_state) => {
+        self.handle_search_task(string, search_state, selection_state)
+      },
+      // Unfortunately, due to Rust issue #16223 and the work around we
+      // have in place we need an additional allocation here.
+      event => Some(UiEvent::Custom(Box::new(event)).into()),
     }
   }
 
@@ -181,6 +318,37 @@ impl Handleable for TabBar {
           Key::Char('`') => (None as Option<Event>).maybe_update(self.select_previous(cap)),
           Key::Char('h') => (None as Option<Event>).maybe_update(self.select(-1, cap)),
           Key::Char('l') => (None as Option<Event>).maybe_update(self.select(1, cap)),
+          Key::Char('n') => {
+            let event = match self.search.take() {
+              SearchT::Unset => {
+                self.search = SearchT::Unset;
+
+                let error = InOut::Error("Nothing to search for".to_string());
+                let event = TermUiEvent::SetInOut(error);
+                UiEvent::Custom(Box::new(event)).into()
+              },
+              SearchT::Taken => panic!("invalid search state"),
+              SearchT::State(string, search_state, mut selection_state) => {
+                let iter = self.tabs.iter().map(|x| x.1);
+                let new_idx = selection_state.normalize(iter);
+                let tab = self.tabs.get(new_idx).unwrap().1;
+
+                let event1 = TermUiEvent::SetInOut(InOut::Search(string.clone()));
+                let event1 = UiEvent::Custom(Box::new(event1));
+
+                let event2 = TermUiEvent::SearchTask(string, search_state, selection_state);
+                let event2 = UiEvent::Returnable(self.id, tab, Box::new(event2));
+
+                UiEvents::from(event1).chain(event2)
+              },
+            };
+            Some(event)
+          },
+          Key::Char('/') => {
+            let event = TermUiEvent::SetInOut(InOut::Input("".to_string(), 0));
+            let event = UiEvent::Custom(Box::new(event));
+            Some(event.into())
+          },
           _ => Some(event.into()),
         }
       },
@@ -193,5 +361,38 @@ impl Handleable for TabBar {
       Ok(e) => self.handle_custom_event(e, cap),
       Err(e) => panic!("Received unexpected custom event: {:?}", e),
     }
+  }
+}
+
+
+#[cfg(test)]
+mod tests {
+  use super::*;
+
+  type TestSelectionState = SelectionStateT<u16>;
+
+
+  #[test]
+  fn search_take() {
+    let mut search = SearchT::Unset::<u16>;
+    assert_eq!(search.take(), SearchT::Unset);
+    assert_eq!(search, SearchT::Taken);
+
+    let mut search = SearchT::Taken::<u16>;
+    assert_eq!(search.take(), SearchT::Taken);
+    assert_eq!(search, SearchT::Taken);
+
+    let selection_state = TestSelectionState::new(42);
+    let mut search = SearchT::State("test".to_string(), SearchState::First, selection_state);
+
+    match search.take() {
+      SearchT::State(string, search_state, selection_state) => {
+        assert_eq!(string, "test");
+        assert_eq!(search_state, SearchState::First);
+        assert_eq!(selection_state, TestSelectionState::new(42));
+      },
+      _ => assert!(false),
+    }
+    assert_eq!(search, SearchT::Taken);
   }
 }